complete refund money method

+++ backend/internal/domain/port/payment_gateway_interface.go
@@ -10,5 +10,5 @@ import (
 type PaymentGateway interface {
 	Create(ctx context.Context, amount int64, returnURL string, orderID uuid.UUID) (string, string, error)
 	GetStatus(ctx context.Context, externalID string) (model.TransactionStatus, error)
-	Refund(ctx context.Context, externalID string, amount float64, idempotencyKey string) error
+	Refund(ctx context.Context, externalID string, amount int64, idempotencyKey string) error
 }
